requirements.txt
socket
os
pybase64
time
csv
datetime
Tkinter
pycryptodome
art
pwntools
getpass